refactor(client): split voice_lounge_screen.dart into focused widgets
